Renamed EventList to AllEventList for semantics and to prevent name collision.  Refactor patient dashboard
