+++ App/components/AllEventList/styles.js
