+++ App/components/AllEventList/test.js
@@ -1,17 +1,17 @@
 import React, {Component} from 'react';
 import renderer from 'react-test-renderer';
-import EventList from './index.js';
+import AllEventList from './index.js';
 import UserGateway from '../../data/UserGateway';
 import EventGateway from '../../data/EventGateway';
 
-describe('EventList', () => {
+describe('AllEventList', () => {
     beforeEach(() => {
         UserGateway.login('sarah@emailprovider.com'); //todo: mock/stub gateways
     });
 
     it('should render correctly', () => {
         const list = renderer.create(
-            <EventList />
+            <AllEventList />
         )
 
         expect(list).toMatchSnapshot();
@@ -21,7 +21,7 @@ describe('EventList', () => {
         const spy = jest.spyOn(EventGateway, 'subscribe');
 
         const list = renderer
-            .create(<EventList />)
+            .create(<AllEventList />)
             .getInstance();
 
         expect(spy).toHaveBeenCalledWith(list.updateEvents);
@@ -31,7 +31,7 @@ describe('EventList', () => {
         const spy = jest.spyOn(EventGateway, 'getAll');
 
         const list = renderer
-            .create(<EventList />)
+            .create(<AllEventList />)
             .getInstance();
 
         expect(spy).toHaveBeenCalled();
@@ -42,7 +42,7 @@ describe('EventList', () => {
         const events = EventGateway.getAll(); //todo: stub this out
 
         const list = renderer
-            .create(<EventList />)
+            .create(<AllEventList />)
             .getInstance();
 
         events.forEach((event) => expect(spy).toHaveBeenCalledWith(event.volunteerId));
@@ -52,7 +52,7 @@ describe('EventList', () => {
       it('should create an event through the gateway', () => {
         const spy = jest.spyOn(EventGateway, 'createNewEvent');
         const list = renderer
-            .create(<EventList />)
+            .create(<AllEventList />)
             .getInstance();
 
           list._onCreateEvent({some: 'data'});
